requirements.txt
# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiofiles==23.2.1

# Data
yfinance==1.2.0
pandas==2.2.2
pydantic==2.7.4

# Database
sqlalchemy==2.0.30

# Environment
python-dotenv==1.0.1

# Robinhood
robin_stocks==3.0.6
python-dateutil==2.9.0

# AI chat
anthropic==0.25.0

# Testing
pytest==8.2.0
httpx==0.27.0
